moved icon start/end to RouteLayer, updated tests
